requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
appdirs==1.4.4
asteroid==0.5.2
asteroid-filterbanks==0.4.0
async-timeout==3.0.1
attrs==21.2.0
audioread==2.1.9
#backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
boto3==1.19.2
botocore==1.22.2
Bottleneck==1.3.2
bravado==11.0.3
bravado-core==5.17.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.5
click==8.0.3
cycler==0.10.0
#decorator @ file:///tmp/build/80754af9/decorator_1632776554403/work
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.3.2
filelock==3.0.12
fonttools==4.25.0
fsspec==2021.8.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.40.0
huggingface-hub==0.0.17
idna==3.2
imageio==2.16.0
importlib-metadata==4.8.1
intervaltree==3.1.0
#ipython @ file:///tmp/build/80754af9/ipython_1632141056704/work
jams==0.3.4
#jedi @ file:///tmp/build/80754af9/jedi_1611333758854/work
jmespath==0.10.0
joblib==1.1.0
jsonpointer==2.1
jsonref==0.2
jsonschema==3.2.0
julius==0.2.5
#kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612282414123/work
librosa==0.8.1
llvmlite==0.37.0
Markdown==3.3.4
#matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1634667019719/work
#matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mir-eval==0.6
mkl-fft
#mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
monotonic==1.6
msgpack==1.0.2
multidict==5.1.0
munkres==1.1.4
neptune-client==0.16.3
neptune-pytorch-lightning==0.9.7
networkx==2.6.3
numba==0.54.1
#numexpr @ file:///tmp/build/80754af9/numexpr_1618856527787/work
numpy==1.20.0
oauthlib==3.1.1
olefile==0.46
packaging==21.0
pandas==1.3.4
#parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pathtools==0.1.2
pb-bss-eval==0.0.2
pesq==0.0.3
#pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
#pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.1
pooch==1.5.2
promise==2.3
#prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.17.3
psutil==5.8.0
#ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyDeprecate==0.3.1
#Pygments @ file:///tmp/build/80754af9/pygments_1629234116488/work
PyJWT==2.3.0
#pymtg @ git+https://github.com/MTG/pymtg@f3f13480a3f8fc5bfafad7611f20940cfc12dbdf
#pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyrsistent==0.18.0
pystoi==0.3.3
#python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytorch-lightning==1.4.9
pytorch-ranger==0.1.1
pytz==2021.3
PyWavelets==1.2.0
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rfc3987==1.3.8
rsa==4.7.2
s3transfer==0.5.0
scikit-image==0.19.1
scikit-learn==1.0.1
#scipy @ file:///tmp/build/80754af9/scipy_1630606797510/work
#seaborn @ file:///tmp/build/80754af9/seaborn_1629307859561/work
sentry-sdk==1.5.12
setproctitle==1.2.3
shortuuid==1.0.9
simplejson==3.17.5
#six @ file:///tmp/build/80754af9/six_1623709665295/work
sklearn==0.0
smmap==5.0.0
sortedcontainers==2.4.0
soundata==0.1.1
SoundFile==0.10.3.post1
strict-rfc3339==0.7
swagger-spec-validator==2.7.4
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==3.0.0
tifffile==2021.11.2
torch==1.9.0
torch-optimizer==0.1.0
torch-stoi==0.1.2
torchaudio==0.9.0
torchmetrics==0.5.1
torchvision==0.10.0
#tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
tqdm==4.62.2
#traitlets @ file:///tmp/build/80754af9/traitlets_1632522747050/work
#typing-extensions @ file:///Users/ktietz/demo/mc3/conda-bld/typing_extensions_1629887998905/work
urllib3==1.26.6
wandb==0.12.18
#wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webcolors==1.11.1
websocket-client==1.2.1
Werkzeug==2.0.1
yarl==1.6.3
zipp==3.5.0
